Properly loading dependencies
