--- domain/lib/mumuki/domain/mumukit/platform.rb
@@ -1,6 +1,0 @@
-require 'mumukit/platform'
-
-Mumukit::Platform.configure do |config|
-  config.user_class_name = 'User'
-  config.organization_class_name = 'Organization'
-end
