+++ lib/mumuki/laboratory.rb
@@ -9,21 +9,41 @@ module Mumuki
   end
 end
 
-require 'mumukit/assistant'
-require 'mumukit/randomizer'
-require 'mumukit/inspection'
-require 'mumukit/bridge'
-require 'mumukit/content_type'
-require 'mumukit/directives'
 require 'mumuki/domain'
+require 'mumukit/login'
+require 'mumukit/nuntius'
+require 'mumukit/platform'
 
 require 'kaminari'
 require 'bootstrap-kaminari-views'
 
-require_relative './laboratory/mumukit/auth'
-require_relative './laboratory/mumukit/login'
-require_relative './laboratory/mumukit/nuntius'
-require_relative './laboratory/mumukit/platform'
+Mumukit::Nuntius.configure do |config|
+  config.app_name = 'laboratory'
+end
+
+Mumukit::Platform.configure do |config|
+  config.application = Mumukit::Platform.laboratory
+  config.web_framework = Mumukit::Platform::WebFramework::Rails
+end
+
+class Mumuki::Laboratory::Engine < ::Rails::Engine
+  config.i18n.available_locales = Mumukit::Platform::Locale.supported
+end
+
+module Mumukit::Platform::OrganizationMapping::Path
+  class << self
+    alias __organization_name__ organization_name
+
+    def organization_name(request, domain)
+      name = __organization_name__(request, domain)
+      if %w(auth login logout).include? name
+        'central'
+      else
+        name
+      end
+    end
+  end
+end
 
 require_relative './laboratory/extensions'
 require_relative './laboratory/controllers'
