--- lib/mumuki/laboratory/mumukit/auth.rb
@@ -1,5 +1,0 @@
-require 'mumukit/auth'
-
-Mumukit::Auth.configure do |config|
-
-end
