--- lib/mumuki/laboratory/mumukit/login.rb
@@ -1,5 +1,0 @@
-require 'mumukit/login'
-
-Mumukit::Login.configure do |config|
-
-end
