--- lib/mumuki/laboratory/mumukit/nuntius.rb
@@ -1,5 +1,0 @@
-require 'mumukit/nuntius'
-
-Mumukit::Nuntius.configure do |config|
-  config.app_name = 'laboratory'
-end
